requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.3
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.4
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.1
importlib-resources==5.2.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.1.0
psycopg2-binary==2.9.1
python-dotenv==0.17.1
slackclient==2.9.3
SQLAlchemy==1.4.25
typing-extensions==3.10.0.0
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.5.0
